tried a kodi version sensing algorithm
Tried to use the same script for both v.18 and v.19 of Kodi, Right now, while OSMC is testing v.19, there seems to be a similar issue, as in Leia nightly, which is no automated dependencies installation. So I chose to do a scripted install of current addon dependencies, for the Netflix addon.